+++ rust/cedar-policy-mcp-schema-generator/src/generator.rs
@@ -1,5 +1,8 @@
 mod err;
+mod identifiers;
+mod request;
 mod schema;
 
-pub use err::SchemaGeneratorError;
+pub use err::{RequestGeneratorError, SchemaGeneratorError};
+pub use request::RequestGenerator;
 pub use schema::{SchemaGenerator, SchemaGeneratorConfig};