+++ rust/mcp-tools-sdk/src/data.rs
@@ -78,11 +78,17 @@ pub enum TypedValue {
     Array(Vec<TypedValue>),
     Tuple(Vec<TypedValue>),
     Union {
-        // index of which type within union type this validates against
+        /// index of which type within union type this validates against
         index: usize,
         value: Box<TypedValue>,
     },
-    Object(HashMap<SmolStr, TypedValue>),
+    Object {
+        /// properties that were explicitly defined by object's type schema
+        properties: HashMap<SmolStr, TypedValue>,
+        /// properties that were not explicitly defined by object's type schema
+        /// but matched the object's `additionalProperty` type
+        additional_properties: HashMap<SmolStr, TypedValue>,
+    },
     Ref {
         name: SmolStr,
         val: Box<TypedValue>,