+++ rust/cedar-policy-mcp-schema-generator/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Adds the ability to convert a MCP tool request (and optionally response) to a Cedar request and entity data compliant to the generated Schema.
+
 ### Fixed
 
 - Generated JSON format schemas now correctly reference common types using `EntityOrCommon` instead `Entity`.